Basic skeleton
I need to group the audit data on UUID (or path, but the problem is the
same). Therefore I don’t have the file name yet when I get the
aggregate data from the audit. So I need to convert the UUID into a
file name (and maybe other stuff…path?). This is the first working
attempt at doing that.